# api/openapi.yaml
openapi: "3.0.3"
info:
  title: ReMail API
  description: |
    ReMail 邮箱交易平台 — 统一业务 API。
    控制台和 SDK 共用同一接口，由鉴权中间件决定调用者能力。
  version: "0.1.0"
  contact:
    name: ReMail Team
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /v1/resources:
    get:
      operationId: getResources
      summary: List email resources
      tags:
        - Resources
      security:
        - cookieAuth: []
      parameters:
        - name: scope
          in: query
          schema:
            type: string
            default: owned
        - name: type
          in: query
          schema:
            type: string
            default: all
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 10000
      responses:
        "200":
          description: Resource list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceListResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/resources/{resourceId}:
    get:
      operationId: getResourceDetail
      summary: Get resource detail (no credentials)
      tags:
        - Resources
      security:
        - cookieAuth: []
      parameters:
        - name: resourceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resource detail
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/MicrosoftResourceDetail"
                  - $ref: "#/components/schemas/DomainResourceDetail"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteResource
      summary: Delete an owned private resource
      tags:
        - Resources
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - name: resourceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Private resource deleted
        "400":
          description: Invalid resource ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Resource is not private or is already deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/resources/publish:
    post:
      operationId: postResourcePublishBatch
      summary: Publish owned resources to public supply
      tags:
        - Resources
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishResourcesRequest"
      responses:
        "200":
          description: Resources published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishResourcesResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Invalid resource status/filter or resource is not private
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/resources/delete:
    post:
      operationId: postResourceDeleteBatch
      summary: Delete owned private resources
      tags:
        - Resources
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteResourcesRequest"
      responses:
        "200":
          description: Resources deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResourcesResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Invalid resource status/filter or resource is not private
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/resources/{resourceId}/publish:
    post:
      operationId: postResourcePublish
      summary: Publish an owned resource to public supply
      tags:
        - Resources
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - name: resourceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resource published
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/MicrosoftResourceDetail"
                  - $ref: "#/components/schemas/DomainResourceDetail"
        "400":
          description: Invalid resource ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Invalid resource status or resource is not private
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /v1/resource-imports/{importId}:
    get:
      operationId: getResourceImport
      summary: Get resource import status
      tags:
        - Resources
      security:
        - cookieAuth: []
      parameters:
        - name: importId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resource import status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportStatusResponse"
        "400":
          description: Invalid import id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Import not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
      description: CSRF token from the csrf_token SameSite cookie; required for authenticated state-changing requests.
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sid
      description: HttpOnly session cookie set after login
  schemas:
...
    Error:
      type: object
      required:
        - message
        - requestId
      properties:
        message:
          type: string
          description: Safe business error message for the user
          example: "Account or password is incorrect."
        requestId:
          type: string
          description: Correlation ID for troubleshooting
          example: "req_a1b2c3d4"
        fields:
          type: object
          description: Field-level validation errors (optional)
          additionalProperties:
            type: string
            example: "Email format is invalid."
      additionalProperties: false

    # --- IAM Schemas ---
...
    ResourceListResponse:
      type: object
      required:
        - items
        - total
        - offset
        - limit
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ResourceItem"
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
      additionalProperties: false
...
    MicrosoftResourceDetail:
      type: object
      required:
        - id
        - emailAddress
        - forSale
        - longLived
        - graphAvailable
        - status
        - qualityScore
        - createdAt
      properties:
        id:
          type: integer
        emailAddress:
          type: string
        forSale:
          type: boolean
        longLived:
          type: boolean
        graphAvailable:
          type: boolean
        status:
          type: string
        qualityScore:
          type: integer
        lastSafeError:
          type: string
        lastAllocatedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      additionalProperties: false
...
    PublishResourcesRequest:
      type: object
      required:
        - selection
      properties:
        selection:
          $ref: "#/components/schemas/ResourceBulkSelection"
      additionalProperties: false
...
    DeleteResourcesRequest:
      type: object
      required:
        - selection
      properties:
        selection:
          $ref: "#/components/schemas/ResourceBulkSelection"
      additionalProperties: false
...